+++ piyavkin.anton/P4/downtrimatrix.hpp
@@ -1,0 +1,5 @@
+#ifndef DOWNTRIMATRIX_H
+#define DOWNTRIMATRIX_H
+#include <cstddef>
+bool DownTriMatrix(size_t result, int rows, int cols, int * matrix);
+#endif
