--- piyavkin.anton/P4/input.txt
@@ -1,1 +1,0 @@
-1 2 3 4 5 6 7 8 9
