+++ piyavkin.anton/P4/inputarray.hpp
@@ -1,0 +1,5 @@
+#ifndef INPUTARRAY_H
+#define INPUTARRAY_H
+#include <iostream>
+size_t InputArray(std::istream & in, int * a, size_t s, size_t toread);
+#endif
