+++ piyavkin.anton/P4/nonzero.cpp
@@ -1,0 +1,13 @@
+#include "nonzero.hpp"
+
+bool NonZero(size_t result, int * matrix)
+{
+  for (size_t i = 0; i < result; ++i)
+  {
+    if (matrix[i] != 0)
+    {
+      return true;
+    }
+  }
+  return false;
+}
