+++ piyavkin.anton/P4/nonzero.hpp
@@ -1,0 +1,5 @@
+#ifndef NONZERO_H
+#define NONZERO_H
+#include <cstddef>
+bool NonZero(size_t result, int * matrix);
+#endif
